+++ Server/src/Server.java
@@ -1,11 +1,10 @@
-import Communication.Helpers.Auxiliary;
 import Communication.Links.LinkMessages.Base.Contracts.ILinkMessage;
 import Communication.Links.LinkMessages.AckMessage;
 import Communication.Links.AuthenticatedPerfectLink;
 import Communication.Links.Data.MessageDeliveryTuple;
 import Communication.Links.LinkMessages.Base.LinkMessageType;
 import Communication.Links.LinkMessages.UdpLinkMessage;
-import Communication.Security.DigitalSignatureAuth;
+import Communication.Links.Security.DigitalSignatureAuth;
 
 import java.net.SocketException;
 import java.security.NoSuchAlgorithmException;
